backend/requirements.txt
# ─────────────────────────────────────────────────────────────────────────
# SmartAgri Unified Backend Requirements
# Merged from:
# - backend/requirements.txt (core FastAPI)
# - farm_advisory/requirements.txt (voice service)
# ─────────────────────────────────────────────────────────────────────────

# ──────────────────── CORE FASTAPI ───────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# ──────────────────── DATABASE ────────────────────────────────────
motor==3.5.0
pymongo[srv]==4.8.0
dnspython==2.6.1

# ──────────────────── AUTHENTICATION ──────────────────────────────
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0

# ──────────────────── SCHEDULING ─────────────────────────────────
apscheduler==3.10.4

# ──────────────────── AI SERVICES ────────────────────────────────
google-generativeai==0.8.0
groq==0.4.1

# ──────────────────── HTTP & NETWORKING ─────────────────────────
requests==2.31.0
httpx==0.27.0
aiosmtplib==3.0.1

# ──────────────────── CONFIGURATION ──────────────────────────────
python-dotenv==1.0.1

# ──────────────────── VOICE & AUDIO ──────────────────────────────
# Speech recognition from WebM/MP4 audio
SpeechRecognition==3.10.1
# Audio format conversion
pydub==0.25.1
# Audio playback (Windows/Linux/Mac)
pyaudio==0.2.14
# Text-to-speech synthesis
gTTS==2.4.0
# Text processing for NLTK-based BLEU scoring
nltk==3.8.1

# ─────────────────────────────────────────────────────────────────────────
# NOTES:
# - Flask/flask-cors removed: FastAPI serves all HTTP routes
# - keyboard removed: not needed in server environments
# - pyaudio included: optional on server (graceful fallback if missing)
# - ffmpeg: install separately (system package):
#   Ubuntu: sudo apt-get install ffmpeg
#   Mac: brew install ffmpeg
#   Windows: choco install ffmpeg
# ─────────────────────────────────────────────────────────────────────────
